Big commit with much new functionality
A lot of work done in exploratory manner.

- Update README.md with links to the reference of VAX BASIC
- Add basic .gitignore

Adds new interactive interpreter mode (cli.js)

Adds Context class to hold the registers and variables
of the interpreter

Implements evaluation of the following statements:

- RUN
- GOTO
- PRINT

The RUN command does not work currently, as it does not
proceed to the next line.

A lot of code formatting changes and a few variable name
changes.

A Token class is added to be used instead of anonymous
objects.

As BASIC (this dialect at least) is very line centric,
a Line class has been added that holds the line number
and a list of statements.

Adds support for parsing LIST, RUN and empty statements.

Improved parsing of PRINT, GOTO

Adds new Statement class to be used instead of anonymous
objects.

Adds Program class that mainly takes care of inserting,
replacing and deleting lines, plus map between line
number and index in the lines array.