Add parsing of expressions (#1)
The primary work of this commit is to implement parsing of expressions. A shunting yard algorithm
is used, extended with support for functions with variable argument count. The output is an
expression tree and not a postfix representation of the expression as is otherwise typical for
the algorithm.

A number of unrelated changes is also included in this commit:

- Setup VS Code launchers, to enable debugging of tests 
- Add basic VS Code workspace settings
- Minor updates in the README
- Fix `BasicArray` implementation. It now uses a single array as data store.
- Fix one error in `BasicArray` test
- Add constants to `Context`
- Add output stream to `Context`
- Methods to assign constants and variables and get value by name in `Context`
- The PRINT and LIST statements now prints to the context stdout, which defaults to `process.stdout`
- Basic support for Functions, and an implementation of the builtin `sin` function
- Add integration tests, that runs BASIC code and validates the output
- Add Enum util to build enums, as there's no builtin support for that in plain JavaScript