Add stream system (#5)
To allow Bajsic to be used in more places than just the Node console, a very simple stream
system is used. This allows to use the interpreter in for example a web application with no
stdin/stdout.

This change made the problem with code being evaluated in many places even more obvious,
so I've replaced more or less all the different locations where code was evaluated to a single
spot, in `eval.js`.

Readline unfortunately can't be used with streams. Hopefully I will be able to add readline again
or similar functionality in the future.

Other changes:

- Adds TypeScript types.
- Program loading has moved from `setupEnvironment` to `Program.loadFromString`