Add support for READ and DATA statements (#14)
This commit adds support for DATA and READ statements, which were only implemented as stubs.
When starting program execution, the context is initialized with data from all DATA statements
and the data index pointer is set to zero. For each data point read, using `Context.getData()`,
the index pointer is increased. If no data is available `OutOfDataError` is thrown.

Also:

- Enforce type checking VS Code using JSDoc
- Integration tests can expect errors using this syntax: `--- throws: "error message" ---`
- Adds basic support for float values
- Helper function `valueFromToken` added to `parser/utils`