Improved handling of program counter (#23)
This commit flattens all statements and put them in one big "statements" array in the program. This enables the program counter (`context.pc`) to point at a specific statement, instead of a line, which solves the problem with multiple statements on a single line where one of them is a gosub or similar.

To accomplish this, the IF statement is translated into a new, simpler statement: `BranchStatement`. It will jump to one of two possible labels (then- and else-label), and at the end of the then-statement block an extra goto is inserted that jumps to the end of the else block. This is all done in `program.flatten` which is called as the last step by the parser.

Fixes #11. Fixes #19.

This commit also implements the `SPACE$` function. Fixes #22.